+++ src/sketch_framework/univmon.rs
@@ -26,7 +26,7 @@
 use crate::common::heap::HHHeap;
 use crate::common::{BOTTOM_LAYER_FINDER, DataInput, hash_item64_seeded, hash64_seeded};
 use crate::common::{L2HH, Vector1D};
-use crate::sketches::count::CountL2HH;
+use crate::sketches::count_topk::CountL2HH;
 use rmp_serde::{
     decode::Error as RmpDecodeError, encode::Error as RmpEncodeError, from_slice, to_vec_named,
 };
