+++ src/sketches/count_topk.rs
@@ -1,17 +1,27 @@
-//! CSHeap: a convenient wrapper that pairs a [`Count`] sketch with an
-//! [`HHHeap`] for automatic top-k heavy-hitter tracking.
+//! Top-k heavy-hitter variants of the Count Sketch family.
 //!
-//! Every insertion updates both the frequency sketch and the heap, mirroring
-//! the pattern used by `CMSHeap` but with Count Sketch (median estimator).
+//! - [`CSHeap`]: pairs a [`Count`] sketch with an [`HHHeap`] for automatic top-k tracking.
+//!   Every insertion updates both the frequency sketch and the heap, mirroring
+//!   the pattern used by `CMSHeap` but with Count Sketch (median estimator).
+//! - [`CountL2HH`]: Count Sketch augmented with per-row L2 norm tracking for
+//!   heavy-hitter detection.
+
+use rmp_serde::{
+    decode::Error as RmpDecodeError, encode::Error as RmpEncodeError, from_slice, to_vec_named,
+};
+use serde::{Deserialize, Serialize};
+use std::marker::PhantomData;
 
 use crate::sketches::count::CountSketchCounter;
 use crate::{
     Count, DataInput, DefaultMatrixI32, DefaultMatrixI64, DefaultMatrixI128, DefaultXxHasher,
     FastPath, FixedMatrix, HHHeap, MatrixStorage, QuickMatrixI64, QuickMatrixI128, RegularPath,
-    SketchHasher, Vector2D, heap_item_to_sketch_input,
+    SketchHasher, Vector1D, Vector2D, compute_median_inline_f64, heap_item_to_sketch_input,
 };
 
 const DEFAULT_TOP_K: usize = 32;
+const DEFAULT_ROW_NUM: usize = 3;
+const DEFAULT_COL_NUM: usize = 4096;
 
 /// A Count Sketch paired with a top-k heavy-hitter heap.
 ///
@@ -889,6 +899,286 @@ mod tests {
         assert!(
             (overlap as f64) / (top_k as f64) >= 0.8,
             "heap overlap too low: overlap={overlap}, regular={regular_heap_keys:?}, fast={fast_heap_keys:?}"
+        );
+    }
+}
+
+// =====================================================================
+// CountL2HH — Count Sketch with per-row L2 norm tracking for
+// heavy-hitter detection. Moved here from `count.rs` so the algorithm
+// file stays focused on plain Count Sketch.
+// =====================================================================
+
+/// Count Sketch augmented with per-row L2 norm tracking for heavy-hitter detection.
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(bound = "")]
+pub struct CountL2HH<H: SketchHasher = DefaultXxHasher> {
+    counts: Vector2D<i64>,
+    l2: Vector1D<i64>,
+    row: usize,
+    col: usize,
+    seed_idx: usize,
+    #[serde(skip)]
+    _hasher: PhantomData<H>,
+}
+
+// Default CountL2HH configuration.
+impl Default for CountL2HH {
+    fn default() -> Self {
+        Self::with_dimensions(DEFAULT_ROW_NUM, DEFAULT_COL_NUM)
+    }
+}
+
+// CountL2HH constructors and operations.
+impl<H: SketchHasher> CountL2HH<H> {
+    /// Creates a sketch with the requested number of rows and columns.
+    pub fn with_dimensions(rows: usize, cols: usize) -> Self {
+        Self::with_dimensions_and_seed(rows, cols, 0)
+    }
+
+    /// Creates a sketch with the requested dimensions and a custom seed offset.
+    pub fn with_dimensions_and_seed(rows: usize, cols: usize, seed_idx: usize) -> Self {
+        let mut sk = CountL2HH {
+            counts: Vector2D::init(rows, cols),
+            l2: Vector1D::filled(rows, 0),
+            row: rows,
+            col: cols,
+            seed_idx,
+            _hasher: PhantomData,
+        };
+        sk.counts.fill(0);
+        sk
+    }
+
+    /// Number of rows in the sketch.
+    pub fn rows(&self) -> usize {
+        self.row
+    }
+
+    /// Number of columns in the sketch.
+    pub fn cols(&self) -> usize {
+        self.col
+    }
+
+    /// Exposes the backing matrix for inspection/testing.
+    pub fn as_storage(&self) -> &Vector2D<i64> {
+        &self.counts
+    }
+
+    /// Mutable access used internally for testing scenarios.
+    pub fn as_storage_mut(&mut self) -> &mut Vector2D<i64> {
+        &mut self.counts
+    }
+
+    /// Merges another sketch while asserting compatible dimensions.
+    pub fn merge(&mut self, other: &Self) {
+        assert_eq!(
+            (self.row, self.col),
+            (other.row, other.col),
+            "dimension mismatch while merging CountL2HH sketches"
+        );
+
+        for i in 0..self.row {
+            for j in 0..self.col {
+                self.counts[i][j] += other.counts[i][j];
+            }
+            self.l2[i] = other.l2[i];
+        }
+    }
+
+    /// Resets all counters and L2 accumulators to zero without reallocating.
+    pub fn clear(&mut self) {
+        self.counts.fill(0);
+        self.l2.fill(0);
+    }
+
+    /// Inserts with hash optimization - computes hash once and reuses it.
+    /// due to the limitation of seeds, use fast_insert only
+    pub fn fast_insert_with_count(&mut self, val: &DataInput, c: i64) {
+        let hashed_val = H::hash128_seeded(self.seed_idx, val);
+        self.fast_insert_with_count_and_hash(hashed_val, c);
+    }
+
+    /// Inserts with hash optimization using precomputed hash value.
+    pub fn fast_insert_with_count_and_hash(&mut self, hashed_val: u128, c: i64) {
+        let mask_bits = self.counts.get_mask_bits() as usize;
+        let mask = (1u128 << mask_bits) - 1;
+        let mut shift_amount = 0;
+        let mut sign_bit_pos = 127;
+
+        for i in 0..self.row {
+            let hashed = (hashed_val >> shift_amount) & mask;
+            let idx = (hashed as usize) % self.col;
+            let bit = ((hashed_val >> sign_bit_pos) & 1) as i64;
+            let sign_bit = -(1 - 2 * bit);
+
+            let old_value = self.counts.query_one_counter(i, idx);
+            let new_value = old_value + sign_bit * c;
+            self.counts[i][idx] = new_value;
+
+            let old_l2 = self.l2.as_slice()[i];
+            let new_l2 = old_l2 + new_value * new_value - old_value * old_value;
+            self.l2[i] = new_l2;
+
+            shift_amount += mask_bits;
+            sign_bit_pos -= 1;
+        }
+    }
+
+    /// Inserts without L2 update using precomputed hash value.
+    pub fn fast_insert_with_count_without_l2_and_hash(&mut self, hashed_val: u128, c: i64) {
+        let mask_bits = self.counts.get_mask_bits() as usize;
+        let mask = (1u128 << mask_bits) - 1;
+        let mut shift_amount = 0;
+        let mut sign_bit_pos = 127;
+
+        for i in 0..self.row {
+            let hashed = (hashed_val >> shift_amount) & mask;
+            let idx = (hashed as usize) % self.col;
+            let bit = ((hashed_val >> sign_bit_pos) & 1) as i64;
+            let sign_bit = -(1 - 2 * bit);
+
+            self.counts[i][idx] += sign_bit * c;
+
+            shift_amount += mask_bits;
+            sign_bit_pos -= 1;
+        }
+    }
+
+    /// Update and estimate with hash optimization.
+    /// due to the limitation of seeds, use fast_insert only
+    pub fn fast_update_and_est(&mut self, val: &DataInput, c: i64) -> f64 {
+        let hashed_val = H::hash128_seeded(self.seed_idx, val);
+        self.fast_insert_with_count_and_hash(hashed_val, c);
+        self.fast_get_est_with_hash(hashed_val)
+    }
+
+    /// Update and estimate without L2 with hash optimization.
+    /// due to the limitation of seeds, use fast_insert only
+    pub fn fast_update_and_est_without_l2(&mut self, val: &DataInput, c: i64) -> f64 {
+        let hashed_val = H::hash128_seeded(self.seed_idx, val);
+        self.fast_insert_with_count_without_l2_and_hash(hashed_val, c);
+        self.fast_get_est_with_hash(hashed_val)
+    }
+
+    /// Returns the median estimate of the squared L2 norm across rows.
+    pub fn get_l2_sqr(&self) -> f64 {
+        let mut values: Vec<f64> = self.l2.as_slice()[..self.row]
+            .iter()
+            .map(|&v| v as f64)
+            .collect();
+        compute_median_inline_f64(&mut values)
+    }
+
+    /// Returns the estimated L2 norm (square root of `get_l2_sqr`).
+    pub fn get_l2(&self) -> f64 {
+        let l2 = self.get_l2_sqr();
+        l2.sqrt()
+    }
+
+    /// Returns the frequency estimate with hash optimization.
+    /// due to the limitation of seeds, use fast_insert only
+    pub fn fast_get_est(&self, val: &DataInput) -> f64 {
+        let hashed_val = H::hash128_seeded(self.seed_idx, val);
+        self.fast_get_est_with_hash(hashed_val)
+    }
+
+    /// Returns the frequency estimate using precomputed hash value.
+    /// due to the limitation of seeds, use fast_insert only
+    pub fn fast_get_est_with_hash(&self, hashed_val: u128) -> f64 {
+        let mask_bits = self.counts.get_mask_bits() as usize;
+        let mask = (1u128 << mask_bits) - 1;
+        let mut lst = Vec::with_capacity(self.row);
+        let mut shift_amount = 0;
+        let mut sign_bit_pos = 127;
+
+        for i in 0..self.row {
+            let hashed = (hashed_val >> shift_amount) & mask;
+            let idx = (hashed as usize) % self.col;
+            let bit = ((hashed_val >> sign_bit_pos) & 1) as i64;
+            let sign_bit = -(1 - 2 * bit);
+            let counter = self.counts.query_one_counter(i, idx);
+            lst.push((sign_bit * counter) as f64);
+
+            shift_amount += mask_bits;
+            sign_bit_pos -= 1;
+        }
+        compute_median_inline_f64(&mut lst[..])
+    }
+
+    /// Serializes the CountL2HH sketch into MessagePack bytes.
+    pub fn serialize_to_bytes(&self) -> Result<Vec<u8>, RmpEncodeError> {
+        to_vec_named(self)
+    }
+
+    /// Deserializes a CountL2HH sketch from MessagePack bytes.
+    pub fn deserialize_from_bytes(bytes: &[u8]) -> Result<Self, RmpDecodeError> {
+        from_slice(bytes)
+    }
+}
+
+#[cfg(test)]
+mod tests_count_l2_hh {
+    use super::*;
+
+    #[test]
+    fn countl2hh_estimates_and_l2_are_consistent() {
+        let mut sketch: CountL2HH = CountL2HH::with_dimensions(3, 32);
+        let key = DataInput::Str("gamma");
+
+        let est_after_first = sketch.fast_update_and_est(&key, 5);
+        assert_eq!(est_after_first, 5.0);
+
+        let est_after_second = sketch.fast_update_and_est(&key, -2);
+        assert_eq!(est_after_second, 3.0);
+
+        let l2 = sketch.get_l2();
+        assert!(l2 >= 3.0, "expected non-trivial l2, got {l2}");
+    }
+
+    #[test]
+    fn countl2hh_merge_combines_frequency_vectors() {
+        let mut left: CountL2HH = CountL2HH::with_dimensions(3, 32);
+        let mut right: CountL2HH = CountL2HH::with_dimensions(3, 32);
+        let key = DataInput::U32(42);
+
+        left.fast_insert_with_count(&key, 4);
+        assert_eq!(left.fast_get_est(&key), 4.0);
+        right.fast_insert_with_count(&key, 9);
+        assert_eq!(right.fast_get_est(&key), 9.0);
+
+        left.merge(&right);
+        assert_eq!(left.fast_get_est(&key), 13.0);
+    }
+
+    #[test]
+    fn countl2hh_round_trip_serialization() {
+        let mut sketch: CountL2HH = CountL2HH::with_dimensions_and_seed(3, 32, 7);
+        let key = DataInput::Str("serialize");
+
+        sketch.fast_insert_with_count(&key, 11);
+        sketch.fast_insert_with_count(&key, -3);
+        let base_est = sketch.fast_get_est(&key);
+        let base_l2 = sketch.get_l2();
+
+        let encoded = sketch
+            .serialize_to_bytes()
+            .expect("serialize CountL2HH into MessagePack");
+        assert!(!encoded.is_empty(), "serialized bytes should not be empty");
+        let data = encoded.clone();
+
+        let decoded: CountL2HH = CountL2HH::deserialize_from_bytes(&data)
+            .expect("deserialize CountL2HH from MessagePack");
+
+        assert_eq!(sketch.rows(), decoded.rows());
+        assert_eq!(sketch.cols(), decoded.cols());
+        assert!(
+            (decoded.fast_get_est(&key) - base_est).abs() < f64::EPSILON,
+            "estimate changed after round trip"
+        );
+        assert!(
+            (decoded.get_l2() - base_l2).abs() < f64::EPSILON,
+            "L2 changed after round trip"
         );
     }
 }
